Introduces abstracttileselectiontool
