+++ src/tiled/magicwandtool.cpp
@@ -25,22 +25,16 @@
 
 #include "brushitem.h"
 #include "tilepainter.h"
-#include "tile.h"
-#include "mapscene.h"
-#include "mapdocument.h"
-#include "changeselectedarea.h"
-
-#include <QApplication>
 
 using namespace Tiled;
 using namespace Tiled::Internal;
 
 MagicWandTool::MagicWandTool(QObject *parent)
-    : AbstractTileTool(tr("Magic Wand"),
-                       QIcon(QLatin1String(
-                               ":images/22x22/stock-tool-fuzzy-select-22.png")),
-                       QKeySequence(tr("W")),
-                       parent)
+    : AbstractTileSelectionTool(tr("Magic Wand"),
+                                QIcon(QLatin1String(
+                                      ":images/22x22/stock-tool-fuzzy-select-22.png")),
+                                QKeySequence(tr("W")),
+                                parent)
 {
 }
 
@@ -52,48 +46,14 @@ void MagicWandTool::tilePositionChanged(const QPoint &tilePos)
         return;
 
     TilePainter regionComputer(mapDocument(), tileLayer);
-    mSelectedRegion = regionComputer.computeFillRegion(tilePos);
-    brushItem()->setTileRegion(mSelectedRegion);
-}
-
-void MagicWandTool::mousePressed(QGraphicsSceneMouseEvent *event)
-{
-    const Qt::MouseButton button = event->button();
-    const Qt::KeyboardModifiers modifiers = event->modifiers();
-
-    if (button != Qt::LeftButton && button != Qt::RightButton)
-        return;
-
-    MapDocument *document = mapDocument();
-
-    QRegion selection;
-
-    // Left button modifies selection, right button clears selection
-    if (button == Qt::LeftButton) {
-        selection = document->selectedArea();
-
-        if (modifiers == Qt::ShiftModifier)
-            selection += mSelectedRegion;
-        else if (modifiers == Qt::ControlModifier)
-            selection -= mSelectedRegion;
-        else if (modifiers == (Qt::ControlModifier | Qt::ShiftModifier))
-            selection &= mSelectedRegion;
-        else
-            selection = mSelectedRegion;
-    }
-
-    if (selection != document->selectedArea()) {
-        QUndoCommand *cmd = new ChangeSelectedArea(document, selection);
-        document->undoStack()->push(cmd);
-    }
-}
-
-void MagicWandTool::mouseReleased(QGraphicsSceneMouseEvent *)
-{
+    setSelectedRegion(regionComputer.computeFillRegion(tilePos));
+    brushItem()->setTileRegion(selectedRegion());
 }
 
 void MagicWandTool::languageChanged()
 {
     setName(tr("Magic Wand"));
     setShortcut(QKeySequence(tr("W")));
+
+    AbstractTileSelectionTool::languageChanged();
 }
