+++ src/tiled/magicwandtool.h
@@ -24,7 +24,7 @@
 #pragma once
 
 
-#include "abstracttiletool.h"
+#include "abstracttileselectiontool.h"
 
 #include "tilelayer.h"
 
@@ -36,24 +36,17 @@ class MapDocument;
 /**
  * Implements a tool that selects a region of connected similar tiles on the layer.
  */
-class MagicWandTool : public AbstractTileTool
+class MagicWandTool : public AbstractTileSelectionTool
 {
     Q_OBJECT
 
 public:
     MagicWandTool(QObject *parent = nullptr);
 
-    void mousePressed(QGraphicsSceneMouseEvent *event) override;
-    void mouseReleased(QGraphicsSceneMouseEvent *event) override;
-
     void languageChanged() override;
 
 protected:
     void tilePositionChanged(const QPoint &tilePos) override;
-
-private:
-
-    QRegion mSelectedRegion;
 };
 
 } // namespace Internal
