+++ src/tiled/selectsametiletool.h
@@ -22,7 +22,7 @@
 #pragma once
 
 
-#include "abstracttiletool.h"
+#include "abstracttileselectiontool.h"
 
 #include "tilelayer.h"
 
@@ -34,24 +34,17 @@ class MapDocument;
 /**
  * Implements a tool that selects a region with all similar tiles on the layer.
  */
-class SelectSameTileTool : public AbstractTileTool
+class SelectSameTileTool : public AbstractTileSelectionTool
 {
     Q_OBJECT
 
 public:
     SelectSameTileTool(QObject *parent = nullptr);
 
-    void mousePressed(QGraphicsSceneMouseEvent *event) override;
-    void mouseReleased(QGraphicsSceneMouseEvent *event) override;
-
     void languageChanged() override;
 
 protected:
     void tilePositionChanged(const QPoint &tilePos) override;
-
-private:
-
-    QRegion mSelectedRegion;
 };
 
 } // namespace Internal
